TINY-7829: moved commands out to separate modules
